# [FIX] analytics: remove partner_id from the name_search