--- a/addons/analytic/models/analytic_account.py
+++ b/addons/analytic/models/analytic_account.py
@@ -12,7 +12,7 @@ class AccountAnalyticAccount(models.Model):
     _description = 'Analytic Account'
     _order = 'plan_id, name asc'
     _check_company_auto = True
-    _rec_names_search = ['name', 'code', 'partner_id']
+    _rec_names_search = ['name', 'code']
 
     name = fields.Char(
         string='Analytic Account',