original commit message backported from https://github.com/odoo/odoo/pull/162951

```txt
Before this commit, our distribution analytics process included searching the table res.partner for potential matches, 
which negatively impacted performance. However, this is no longer necessary as the default name of the account 
matches the name of the partner. This change removes the unnecessary performance overhead.
```


We have reproduced a query that never ends only opening the following report:

```python
report = self.env.ref("account_reports.profit_and_loss")
infos = report.get_report_informations({})
```

Checking the configuration of this report it has domains similar to:
 - `("analytic_line_ids.account_id", "ilike", "VALUE")`

<details>
  <summary>generating a big query running for hours in production</summary>

```sql
-- with partner
SELECT 
  COALESCE(
    SUM(
      ROUND(
        account_move_line.balance * currency_table.rate, 
        currency_table.precision
      )
    ), 
    0.0
  ) AS sum, 
  COUNT(DISTINCT account_move_line.id) AS count_rows 
FROM 
  "account_move_line" 
  LEFT JOIN "account_account" AS "account_move_line__account_id" ON (
    "account_move_line"."account_id" = "account_move_line__account_id"."id"
  ) 
  JOIN (
    VALUES 
      (1, 1.0, 2)
  ) AS currency_table(company_id, rate, precision) ON currency_table.company_id = account_move_line.company_id 
WHERE 
  (
    (
      (
        (
          (
            (
              (
                "account_move_line"."display_type" not in ('line_section', 'line_note')
              ) 
              OR "account_move_line"."display_type" IS NULL
            ) 
            AND (
              "account_move_line"."company_id" in (1)
            )
          ) 
          AND (
            "account_move_line"."date" <= '2024-12-31'
          )
        ) 
        AND (
          "account_move_line"."date" >= '2024-01-01'
        )
      ) 
      AND(
        "account_move_line"."parent_state" = 'posted'
      )
    ) 
    AND (
      (
        (
          "account_move_line__account_id"."account_type" = 'expense'
        ) 
        OR (
          "account_move_line__account_id"."account_type" = 'expense_depreciation'
        )
      ) 
      AND (
        (
          (
            (
              (
                (
                  "account_move_line"."id" IN (
                    SELECT 
                      "account_analytic_line"."move_line_id" 
                    FROM 
                      "account_analytic_line" 
                    WHERE 
                      (
                        "account_analytic_line"."account_id" in (
                          SELECT 
                            "account_analytic_account".id 
                          FROM 
                            "account_analytic_account" 
                            LEFT JOIN "account_analytic_plan" AS "account_analytic_account__plan_id" ON (
                              "account_analytic_account"."plan_id" = "account_analytic_account__plan_id"."id"
                            ) 
                          WHERE 
                            (
                              (
                                (
                                  unaccent(
                                    "account_analytic_account"."name" :: text
                                  ) ilike unaccent('%Consejo%')
                                ) 
                                OR (
                                  unaccent(
                                    "account_analytic_account"."code" :: text
                                  ) ilike unaccent('%Consejo%')
                                )
                              ) 
                              OR (
                                "account_analytic_account"."partner_id" in (
                                  SELECT 
                                    "res_partner".id 
                                  FROM 
                                    "res_partner" 
                                  WHERE 
                                    (
                                      (
                                        (
                                          (
                                            (
                                              unaccent(
                                                "res_partner"."display_name" :: text
                                              ) ilike unaccent('%Consejo%')
                                            ) 
                                            OR (
                                              unaccent("res_partner"."email" :: text) ilike unaccent('%Consejo%')
                                            )
                                          ) 
                                          OR (
                                            unaccent("res_partner"."ref" :: text) ilike unaccent('%Consejo%')
                                          )
                                        ) 
                                        OR (
                                          unaccent("res_partner"."vat" :: text) ilike unaccent('%Consejo%')
                                        )
                                      ) 
                                      OR (
                                        unaccent(
                                          "res_partner"."company_registry" :: text
                                        ) ilike unaccent('%Consejo%')
                                      )
                                    ) 
                                    AND (
                                      (
                                        (
                                          (
                                            "res_partner"."type" != 'private'
                                          ) 
                                          OR "res_partner"."type" IS NULL
                                        ) 
                                        OR "res_partner"."type" IS NULL
                                      ) 
                                      OR ("res_partner"."type" = 'private')
                                    )
                                )
                              )
                            ) 
                            AND (
                              (
                                "account_analytic_account"."company_id" in (1)
                              ) 
                              OR "account_analytic_account"."company_id" IS NULL
                            )
                        )
                      ) 
                      AND (
                        "account_analytic_line"."company_id" in (1)
                      ) 
                      AND "account_analytic_line"."move_line_id" IS NOT NULL
                  )
                ) 
                OR (
                  "account_move_line"."id" IN (
                    SELECT 
                      "account_analytic_line"."move_line_id" 
                    FROM 
                      "account_analytic_line" 
                    WHERE 
                      (
                        "account_analytic_line"."account_id" in (
                          SELECT 
                            "account_analytic_account".id 
                          FROM 
                            "account_analytic_account" 
                            LEFT JOIN "account_analytic_plan" AS "account_analytic_account__plan_id" ON (
                              "account_analytic_account"."plan_id" = "account_analytic_account__plan_id"."id"
                            ) 
                          WHERE 
                            (
                              (
                                (
                                  unaccent(
                                    "account_analytic_account"."name" :: text
                                  ) ilike unaccent('%Finan%')
                                ) 
                                OR (
                                  unaccent(
                                    "account_analytic_account"."code" :: text
                                  ) ilike unaccent('%Finan%')
                                )
                              ) 
                              OR (
                                "account_analytic_account"."partner_id" in (
                                  SELECT 
                                    "res_partner".id 
                                  FROM 
                                    "res_partner" 
                                  WHERE 
                                    (
                                      (
                                        (
                                          (
                                            (
                                              unaccent(
                                                "res_partner"."display_name" :: text
                                              ) ilike unaccent('%Finan%')
                                            ) 
                                            OR (
                                              unaccent("res_partner"."email" :: text) ilike unaccent('%Finan%')
                                            )
                                          ) 
                                          OR (
                                            unaccent("res_partner"."ref" :: text) ilike unaccent('%Finan%')
                                          )
                                        ) 
                                        OR (
                                          unaccent("res_partner"."vat" :: text) ilike unaccent('%Finan%')
                                        )
                                      ) 
                                      OR (
                                        unaccent(
                                          "res_partner"."company_registry" :: text
                                        ) ilike unaccent('%Finan%')
                                      )
                                    ) 
                                    AND (
                                      (
                                        (
                                          (
                                            "res_partner"."type" != 'private'
                                          ) 
                                          OR "res_partner"."type" IS NULL
                                        ) 
                                        OR "res_partner"."type" IS NULL
                                      ) 
                                      OR ("res_partner"."type" = 'private')
                                    )
                                )
                              )
                            ) 
                            AND (
                              (
                                "account_analytic_account"."company_id" in (1)
                              ) 
                              OR "account_analytic_account"."company_id" IS NULL
                            )
                        )
                      ) 
                      AND (
                        "account_analytic_line"."company_id" in (1)
                      ) 
                      AND "account_analytic_line"."move_line_id" IS NOT NULL
                  )
                )
              ) 
              OR (
                "account_move_line"."id" IN (
                  SELECT 
                    "account_analytic_line"."move_line_id" 
                  FROM 
                    "account_analytic_line" 
                  WHERE 
                    (
                      "account_analytic_line"."account_id" in (
                        SELECT 
                          "account_analytic_account".id 
                        FROM 
                          "account_analytic_account" 
                          LEFT JOIN "account_analytic_plan" AS "account_analytic_account__plan_id" ON (
                            "account_analytic_account"."plan_id" = "account_analytic_account__plan_id"."id"
                          ) 
                        WHERE 
                          (
                            (
                              (
                                unaccent(
                                  "account_analytic_account"."name" :: text
                                ) ilike unaccent('%Logis%')
                              ) 
                              OR (
                                unaccent(
                                  "account_analytic_account"."code" :: text
                                ) ilike unaccent('%Logis%')
                              )
                            ) 
                            OR (
                              "account_analytic_account"."partner_id" in (
                                SELECT 
                                  "res_partner".id 
                                FROM 
                                  "res_partner" 
                                WHERE 
                                  (
                                    (
                                      (
                                        (
                                          (
                                            unaccent(
                                              "res_partner"."display_name" :: text
                                            ) ilike unaccent('%Logis%')
                                          ) 
                                          OR (
                                            unaccent("res_partner"."email" :: text) ilike unaccent('%Logis%')
                                          )
                                        ) 
                                        OR (
                                          unaccent("res_partner"."ref" :: text) ilike unaccent('%Logis%')
                                        )
                                      ) 
                                      OR (
                                        unaccent("res_partner"."vat" :: text) ilike unaccent('%Logis%')
                                      )
                                    ) 
                                    OR (
                                      unaccent(
                                        "res_partner"."company_registry" :: text
                                      ) ilike unaccent('%Logis%')
                                    )
                                  ) 
                                  AND (
                                    (
                                      (
                                        (
                                          "res_partner"."type" != 'private'
                                        ) 
                                        OR "res_partner"."type" IS NULL
                                      ) 
                                      OR "res_partner"."type" IS NULL
                                    ) 
                                    OR ("res_partner"."type" = 'private')
                                  )
                              )
                            )
                          ) 
                          AND (
                            (
                              "account_analytic_account"."company_id" in (1)
                            ) 
                            OR "account_analytic_account"."company_id" IS NULL
                          )
                      )
                    ) 
                    AND (
                      "account_analytic_line"."company_id" in (1)
                    ) 
                    AND "account_analytic_line"."move_line_id" IS NOT NULL
                )
              )
            ) 
            OR (
              "account_move_line"."id" IN (
                SELECT 
                  "account_analytic_line"."move_line_id" 
                FROM 
                  "account_analytic_line" 
                WHERE 
                  (
                    "account_analytic_line"."account_id" in (
                      SELECT 
                        "account_analytic_account".id 
                      FROM 
                        "account_analytic_account" 
                        LEFT JOIN "account_analytic_plan" AS "account_analytic_account__plan_id" ON (
                          "account_analytic_account"."plan_id" = "account_analytic_account__plan_id"."id"
                        ) 
                      WHERE 
                        (
                          (
                            (
                              unaccent(
                                "account_analytic_account"."name" :: text
                              ) ilike unaccent('%Recur%')
                            ) 
                            OR (
                              unaccent(
                                "account_analytic_account"."code" :: text
                              ) ilike unaccent('%Recur%')
                            )
                          ) 
                          OR (
                            "account_analytic_account"."partner_id" in (
                              SELECT 
                                "res_partner".id 
                              FROM 
                                "res_partner" 
                              WHERE 
                                (
                                  (
                                    (
                                      (
                                        (
                                          unaccent(
                                            "res_partner"."display_name" :: text
                                          ) ilike unaccent('%Recur%')
                                        ) 
                                        OR (
                                          unaccent("res_partner"."email" :: text) ilike unaccent('%Recur%')
                                        )
                                      ) 
                                      OR (
                                        unaccent("res_partner"."ref" :: text) ilike unaccent('%Recur%')
                                      )
                                    ) 
                                    OR (
                                      unaccent("res_partner"."vat" :: text) ilike unaccent('%Recur%')
                                    )
                                  ) 
                                  OR (
                                    unaccent(
                                      "res_partner"."company_registry" :: text
                                    ) ilike unaccent('%Recur%')
                                  )
                                ) 
                                AND (
                                  (
                                    (
                                      (
                                        "res_partner"."type" != 'private'
                                      ) 
                                      OR "res_partner"."type" IS NULL
                                    ) 
                                    OR "res_partner"."type" IS NULL
                                  ) 
                                  OR ("res_partner"."type" = 'private')
                                )
                            )
…