@hupo-odoo 2024-11-28
Thanks for the PR. @h4818 as discussed, this PR looks fine